requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
langgraph>=0.0.19
langchain-core>=0.1.0
requests>=2.31.0
tavily-python>=0.1.9
grandalf>=0.8
